refactors join league screen
